+++ packages/extension-koni-ui/src/util/dom.ts
@@ -4,17 +4,21 @@
 export const waitForElement = (selector: string, callback: (element: Element) => any) => {
   let count = 0;
   const interval = setInterval(function () {
-    const element = document.querySelector(selector);
+    try {
+      const element = document.querySelector(selector);
 
-    if (element) {
-      clearInterval(interval);
-      callback(element);
-    } else {
-      count++;
-
-      if (count >= 10) {
+      if (element) {
         clearInterval(interval);
+        callback(element);
+      } else {
+        count++;
+
+        if (count >= 10) {
+          clearInterval(interval);
+        }
       }
+    } catch (e) {
+      clearInterval(interval);
     }
   }, 100);
 };